+++ http-client/common/src/main/java/dev/rico/samples/http/CityDetails.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.samples.rest;
+package dev.rico.samples.http;
 
 public class CityDetails extends City {
