+++ http-client/server-javaee/src/main/java/dev/rico/samples/http/CityEndpoint.java
@@ -13,24 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.samples.rest;
+package dev.rico.samples.http;
 
-import com.canoo.platform.server.timing.Metric;
-import com.canoo.platform.server.timing.ServerTiming;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
+import dev.rico.server.timing.Metric;
+import dev.rico.server.timing.ServerTiming;
 
-@RestController
-@RequestMapping("/api/city")
+import javax.inject.Inject;
+import javax.ws.rs.GET;
+import javax.ws.rs.POST;
+import javax.ws.rs.Path;
+
+@Path("city")
 public class CityEndpoint {
 
-    @Autowired
+    @Inject
     private ServerTiming serverTiming;
 
-    @RequestMapping(method = RequestMethod.GET)
+    @GET
     public String testMetrics() {
         final Metric metric1 = serverTiming.start("loadCache");
         sleep(1_100);
@@ -47,8 +46,8 @@ public class CityEndpoint {
         return "DONE";
     }
 
-    @RequestMapping(method = RequestMethod.POST)
-    public CityDetails getDetails(@RequestBody final City city) {
+    @POST
+    public CityDetails getDetails(final City city) {
         final Metric metric = serverTiming.start("getCityDetail");
         try {
             final CityDetails cityDetails = new CityDetails(city);
@@ -60,12 +59,11 @@ public class CityEndpoint {
         }
     }
 
-    private void sleep(final long ms) {
+    private void sleep(long ms) {
         try {
             Thread.sleep(ms);
-        } catch (final InterruptedException e) {
+        } catch (InterruptedException e) {
             e.printStackTrace();
         }
     }
-
 }
