+++ http-client/server-spring/src/main/java/dev/rico/samples/http/CityEndpoint.java
@@ -13,23 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.samples.rest;
+package dev.rico.samples.http;
 
-import com.canoo.platform.server.timing.Metric;
-import com.canoo.platform.server.timing.ServerTiming;
+import dev.rico.samples.http.City;
+import dev.rico.samples.http.CityDetails;
+import dev.rico.server.timing.Metric;
+import dev.rico.server.timing.ServerTiming;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
+import org.springframework.web.bind.annotation.RestController;
 
-import javax.inject.Inject;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-
-@Path("city")
+@RestController
+@RequestMapping("/api/city")
 public class CityEndpoint {
 
-    @Inject
+    @Autowired
     private ServerTiming serverTiming;
 
-    @GET
+    @RequestMapping(method = RequestMethod.GET)
     public String testMetrics() {
         final Metric metric1 = serverTiming.start("loadCache");
         sleep(1_100);
@@ -46,8 +49,8 @@ public class CityEndpoint {
         return "DONE";
     }
 
-    @POST
-    public CityDetails getDetails(final City city) {
+    @RequestMapping(method = RequestMethod.POST)
+    public CityDetails getDetails(@RequestBody final City city) {
         final Metric metric = serverTiming.start("getCityDetail");
         try {
             final CityDetails cityDetails = new CityDetails(city);
@@ -59,11 +62,12 @@ public class CityEndpoint {
         }
     }
 
-    private void sleep(long ms) {
+    private void sleep(final long ms) {
         try {
             Thread.sleep(ms);
-        } catch (InterruptedException e) {
+        } catch (final InterruptedException e) {
             e.printStackTrace();
         }
     }
+
 }
