+++ http-client/server-spring/src/main/java/dev/rico/samples/http/ServerApplication.java
@@ -13,37 +13,15 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.samples.rest;
+package dev.rico.samples.http;
 
-import java.io.Serializable;
+import dev.rico.server.RicoApplication;
+import org.springframework.boot.SpringApplication;
 
-public class City implements Serializable {
+@RicoApplication
+public class ServerApplication {
 
-    private String name;
-
-    private String country;
-
-    public City() {
-    }
-
-    public City(String name, String country) {
-        this.name = name;
-        this.country = country;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getCountry() {
-        return country;
-    }
-
-    public void setCountry(String country) {
-        this.country = country;
+    public static void main(final String[] args) {
+        SpringApplication.run(ServerApplication.class);
     }
 }
